+++ NEWS.md
@@ -6,6 +6,7 @@
 
 #### Minor improvements and bug fixes
 * IIASA internal functionalities such as preparation of GLOBIOM data have been transferred to [BNRTools](https://github.com/iiasa/BNRTools)
+* Small bug fixed related to manual provision of scenario thresholds.
 
 # ibis.iSDM 0.1.5